function complete

+++ include/neuron_demo_iic/neuronIIc.hpp
@@ -31,7 +31,7 @@ class NeuronIIc
 		void SetDir(uint32_t dir);
 		void SetLevel(uint32_t level);
 		void ReadLevel(uint32_t& level);
-		void ReadI2C(int16_t *data_ptr, unsigned int bfr_size);
+		void ReadI2C(uint8_t *data_ptr, unsigned int bfr_size);
         void WakeUp6050();
 
         explicit NeuronIIc(uint32_t pin) : pin_(pin) {};
